Refactoring to Win32AppInfo

+++ Microsoft.Toolkit.Uwp.Notifications/Toasts/Compat/ToastNotificationManagerCompat.cs
@@ -12,8 +12,6 @@ using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
 using System.Runtime.InteropServices;
-using System.Security.Cryptography;
-using System.Text;
 using Microsoft.Win32;
 using Windows.ApplicationModel;
 using Windows.Foundation.Collections;
@@ -114,6 +112,8 @@ namespace Microsoft.Toolkit.Uwp.Notifications
                 return;
             }
 
+            Win32AppInfo win32AppInfo = null;
+
             // If sparse
             if (DesktopBridgeHelpers.HasIdentity())
             {
@@ -121,91 +121,25 @@ namespace Microsoft.Toolkit.Uwp.Notifications
             }
             else
             {
-                // Win32 apps are uniquely identified based on their process name
-                _win32Aumid = GetAumidFromCurrentProcess(Process.GetCurrentProcess());
+                win32AppInfo = Win32AppInfo.Get();
+                _win32Aumid = win32AppInfo.Aumid;
             }
 
             // Create and register activator
             var activatorType = CreateAndRegisterActivator();
 
             // Register via registry
-            var currProcess = Process.GetCurrentProcess();
-
             using (var rootKey = Registry.CurrentUser.CreateSubKey(@"Software\Classes\AppUserModelId\" + _win32Aumid))
             {
                 // If they don't have identity, we need to specify the display assets
                 if (!DesktopBridgeHelpers.HasIdentity())
                 {
-                    IShellItem shortcutItem = null;
-                    try
-                    {
-                        IApplicationResolver appResolver = (IApplicationResolver)new CAppResolver();
-                        appResolver.GetBestShortcutForAppID(_win32Aumid, out shortcutItem);
-                    }
-                    catch
-                    {
-                    }
-
-                    string displayName = null;
-                    string iconPath = null;
-
-                    // First we attempt to use display assets from the shortcut itself
-                    if (shortcutItem != null)
-                    {
-                        try
-                        {
-                            shortcutItem.GetDisplayName(0, out displayName);
-
-                            ((IShellItemImageFactory)shortcutItem).GetImage(new SIZE(48, 48), SIIGBF.IconOnly | SIIGBF.BiggerSizeOk, out IntPtr nativeHBitmap);
-
-                            if (nativeHBitmap != IntPtr.Zero)
-                            {
-                                try
-                                {
-                                    System.Drawing.Bitmap bmp = System.Drawing.Bitmap.FromHbitmap(nativeHBitmap);
-
-                                    if (IsAlphaBitmap(bmp, out var bmpData))
-                                    {
-                                        var alphaBitmap = GetAlphaBitmapFromBitmapData(bmpData);
-                                        iconPath = SaveIconToAppPath(alphaBitmap);
-                                    }
-                                    else
-                                    {
-                                        iconPath = SaveIconToAppPath(bmp);
-                                    }
-                                }
-                                catch
-                                {
-                                }
-
-                                DeleteObject(nativeHBitmap);
-                            }
-                        }
-                        catch
-                        {
-                        }
-                    }
-
-                    // If we didn't get a display name from shortcut
-                    if (string.IsNullOrWhiteSpace(displayName))
-                    {
-                        // We use the one from the process
-                        displayName = GetDisplayNameFromCurrentProcess(currProcess);
-                    }
-
-                    // If we didn't get an icon from shortcut
-                    if (string.IsNullOrWhiteSpace(iconPath))
-                    {
-                        // We use the one from the process
-                        iconPath = ExtractAndObtainIconFromCurrentProcess(currProcess);
-                    }
-
                     // Set the display name and icon uri
-                    rootKey.SetValue("DisplayName", displayName);
+                    rootKey.SetValue("DisplayName", win32AppInfo.DisplayName);
 
-                    if (iconPath != null)
+                    if (win32AppInfo.IconPath != null)
                     {
-                        rootKey.SetValue("IconUri", iconPath);
+                        rootKey.SetValue("IconUri", win32AppInfo.IconPath);
                     }
                     else
                     {
@@ -222,135 +156,6 @@ namespace Microsoft.Toolkit.Uwp.Notifications
 
                 rootKey.SetValue("CustomActivator", string.Format("{{{0}}}", activatorType.GUID));
             }
-        }
-
-        // From https://stackoverflow.com/a/9291151
-        private static System.Drawing.Bitmap GetAlphaBitmapFromBitmapData(System.Drawing.Imaging.BitmapData bmpData)
-        {
-            return new System.Drawing.Bitmap(
-                    bmpData.Width,
-                    bmpData.Height,
-                    bmpData.Stride,
-                    System.Drawing.Imaging.PixelFormat.Format32bppArgb,
-                    bmpData.Scan0);
-        }
-
-        // From https://stackoverflow.com/a/9291151
-        private static bool IsAlphaBitmap(System.Drawing.Bitmap bmp, out System.Drawing.Imaging.BitmapData bmpData)
-        {
-            var bmBounds = new System.Drawing.Rectangle(0, 0, bmp.Width, bmp.Height);
-
-            bmpData = bmp.LockBits(bmBounds, System.Drawing.Imaging.ImageLockMode.ReadOnly, bmp.PixelFormat);
-
-            try
-            {
-                for (int y = 0; y <= bmpData.Height - 1; y++)
-                {
-                    for (int x = 0; x <= bmpData.Width - 1; x++)
-                    {
-                        var pixelColor = System.Drawing.Color.FromArgb(
-                            Marshal.ReadInt32(bmpData.Scan0, (bmpData.Stride * y) + (4 * x)));
-
-                        if (pixelColor.A > 0 & pixelColor.A < 255)
-                        {
-                            return true;
-                        }
-                    }
-                }
-            }
-            finally
-            {
-                bmp.UnlockBits(bmpData);
-            }
-
-            return false;
-        }
-
-        /// <summary>Deletes a logical pen, brush, font, bitmap, region, or palette, freeing all system resources associated with the object. After the object is deleted, the specified handle is no longer valid.</summary>
-        /// <param name="hObject">A handle to a logical pen, brush, font, bitmap, region, or palette.</param>
-        /// <returns>
-        ///   <para>If the function succeeds, the return value is nonzero.</para>
-        ///   <para>If the specified handle is not valid or is currently selected into a DC, the return value is zero.</para>
-        /// </returns>
-        /// <remarks>
-        ///   <para>Do not delete a drawing object (pen or brush) while it is still selected into a DC.</para>
-        ///   <para>When a pattern brush is deleted, the bitmap associated with the brush is not deleted. The bitmap must be deleted independently.</para>
-        /// </remarks>
-        [DllImport("gdi32.dll", EntryPoint = "DeleteObject")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        private static extern bool DeleteObject([In] IntPtr hObject);
-
-        private static string GetAumidFromCurrentProcess(Process process)
-        {
-            IApplicationResolver appResolver = (IApplicationResolver)new CAppResolver();
-            appResolver.GetAppIDForProcess(Convert.ToUInt32(process.Id), out string appId, out bool pinningPrevented, out bool explicitAppId, out bool embeddedShortcutValid);
-            return appId;
-        }
-
-        private static string GetDisplayNameFromCurrentProcess(Process process)
-        {
-            // If AssemblyTitle is set, use that
-            var assemblyTitleAttr = Assembly.GetEntryAssembly().GetCustomAttribute<AssemblyTitleAttribute>();
-            if (assemblyTitleAttr != null)
-            {
-                return assemblyTitleAttr.Title;
-            }
-
-            // Otherwise, fall back to process name
-            return process.ProcessName;
-        }
-
-        private static string ExtractAndObtainIconFromCurrentProcess(Process process)
-        {
-            return ExtractAndObtainIconFromPath(process.MainModule.FileName);
-        }
-
-        private static string ExtractAndObtainIconFromPath(string pathToExtract)
-        {
-            try
-            {
-                // Extract the icon
-                var icon = System.Drawing.Icon.ExtractAssociatedIcon(pathToExtract);
-
-                using (var bmp = icon.ToBitmap())
-                {
-                    return SaveIconToAppPath(bmp);
-                }
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        private static string SaveIconToAppPath(System.Drawing.Bitmap bitmap)
-        {
-            try
-            {
-                var path = Path.Combine(GetAppDataFolderPath(), "Icon.png");
-
-                // Ensure the directories exist
-                Directory.CreateDirectory(Path.GetDirectoryName(path));
-
-                bitmap.Save(path, System.Drawing.Imaging.ImageFormat.Png);
-
-                return path;
-            }
-            catch
-            {
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Gets the app data folder path within the ToastNotificationManagerCompat folder, used for storing icon assets and any additional data.
-        /// </summary>
-        /// <returns>Returns a string of the absolute folder path.</returns>
-        private static string GetAppDataFolderPath()
-        {
-            string conciseAumid = _win32Aumid.Contains('\\') ? GenerateGuid(_win32Aumid) : _win32Aumid;
-
-            return Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "ToastNotificationManagerCompat", "Apps", conciseAumid);
         }
 
         private static Type CreateActivatorType()
@@ -376,7 +181,7 @@ namespace Microsoft.Toolkit.Uwp.Notifications
             }
             else
             {
-                _clsid = GenerateGuid(_win32Aumid);
+                _clsid = Win32AppInfo.GenerateGuid(_win32Aumid);
             }
 
             tb.SetCustomAttribute(new CustomAttributeBuilder(
@@ -470,39 +275,6 @@ namespace Microsoft.Toolkit.Uwp.Notifications
             }
 
             return clsid;
-        }
-
-        /// <summary>
-        /// From https://stackoverflow.com/a/41622689/1454643
-        /// Generates Guid based on String. Key assumption for this algorithm is that name is unique (across where it it's being used)
-        /// and if name byte length is less than 16 - it will be fetched directly into guid, if over 16 bytes - then we compute sha-1
-        /// hash from string and then pass it to guid.
-        /// </summary>
-        /// <param name="name">Unique name which is unique across where this guid will be used.</param>
-        /// <returns>For example "706C7567-696E-7300-0000-000000000000" for "plugins"</returns>
-        private static string GenerateGuid(string name)
-        {
-            byte[] buf = Encoding.UTF8.GetBytes(name);
-            byte[] guid = new byte[16];
-            if (buf.Length < 16)
-            {
-                Array.Copy(buf, guid, buf.Length);
-            }
-            else
-            {
-                using (SHA1 sha1 = SHA1.Create())
-                {
-                    byte[] hash = sha1.ComputeHash(buf);
-
-                    // Hash is 20 bytes, but we need 16. We loose some of "uniqueness", but I doubt it will be fatal
-                    Array.Copy(hash, guid, 16);
-                }
-            }
-
-            // Don't use Guid constructor, it tends to swap bytes. We want to preserve original string as hex dump.
-            string guidS = $"{guid[0]:X2}{guid[1]:X2}{guid[2]:X2}{guid[3]:X2}-{guid[4]:X2}{guid[5]:X2}-{guid[6]:X2}{guid[7]:X2}-{guid[8]:X2}{guid[9]:X2}-{guid[10]:X2}{guid[11]:X2}{guid[12]:X2}{guid[13]:X2}{guid[14]:X2}{guid[15]:X2}";
-
-            return guidS;
         }
 
         private static Type CreateAndRegisterActivator()
@@ -729,12 +501,12 @@ namespace Microsoft.Toolkit.Uwp.Notifications
             {
             }
 
-            if (!DesktopBridgeHelpers.HasIdentity())
+            if (!DesktopBridgeHelpers.HasIdentity() && _win32Aumid != null)
             {
                 try
                 {
                     // Delete any of the app files
-                    var appDataFolderPath = GetAppDataFolderPath();
+                    var appDataFolderPath = Win32AppInfo.GetAppDataFolderPath(_win32Aumid);
                     if (Directory.Exists(appDataFolderPath))
                     {
                         Directory.Delete(appDataFolderPath, recursive: true);
